requirements.txt
langchain_core
langchain_community
streamlit
ctransformers
sentence-transformers
langchain
ipykernel
python-dotenv
pypdf
faiss-cpu
-e .